add up ranks totals to get total downloads for user agents
