+++ src/app/ngrx/index.ts
@@ -6,7 +6,7 @@ export { RouterParams, DownloadsTableModel,
   CHARTTYPE_LINE, CHARTTYPE_BAR, CHARTTYPE_HORIZBAR, ChartType,
   METRICSTYPE_DOWNLOADS, METRICSTYPE_DEMOGRAPHICS, METRICSTYPE_TRAFFICSOURCES, MetricsType, getMetricsProperty,
   GROUPTYPE_AGENTNAME, GROUPTYPE_AGENTOS, GROUPTYPE_AGENTTYPE,
-  GROUPTYPE_GEOCOUNTRY, GROUPTYPE_GEOMETRO, GROUPTYPE_GEOSUBDIV, GroupType,
+  GROUPTYPE_GEOCOUNTRY, GROUPTYPE_GEOMETRO, GROUPTYPE_GEOSUBDIV, GroupType, getGroupName,
   Totals, PodcastTotals } from './reducers/models';
 export { PodcastMetricsModel } from './reducers/podcast-metrics.reducer';
 export { EpisodeMetricsModel } from './reducers/episode-metrics.reducer';
