+++ src/app/ngrx/reducers/models/index.ts
@@ -1,7 +1,7 @@
 export { METRICSTYPE_DOWNLOADS, METRICSTYPE_DEMOGRAPHICS, METRICSTYPE_TRAFFICSOURCES,
   MetricsType, getMetricsProperty } from './metrics.type';
 export { GROUPTYPE_AGENTNAME, GROUPTYPE_AGENTOS, GROUPTYPE_AGENTTYPE,
-  GROUPTYPE_GEOCOUNTRY, GROUPTYPE_GEOMETRO, GROUPTYPE_GEOSUBDIV, GroupType } from './group.type';
+  GROUPTYPE_GEOCOUNTRY, GROUPTYPE_GEOMETRO, GROUPTYPE_GEOSUBDIV, GroupType, getGroupName } from './group.type';
 export { INTERVAL_MONTHLY, INTERVAL_WEEKLY, INTERVAL_DAILY, INTERVAL_HOURLY, IntervalModel, IntervalList } from './interval.model';
 export { CHARTTYPE_PODCAST, CHARTTYPE_EPISODES, CHARTTYPE_STACKED,
   CHARTTYPE_LINE, CHARTTYPE_BAR, CHARTTYPE_HORIZBAR, ChartType } from './chart.type';
