+++ src/app/ngrx/reducers/selectors/downloads-summary.selectors.ts
@@ -1,0 +1,19 @@
+import { createSelector } from '@ngrx/store';
+import { selectMetricsTypeRoute } from './router.selectors';
+import { selectPodcastMetricsFilteredTotal } from './podcast-metrics.selectors';
+import { selectRoutedPodcastRanksTotalDownloads } from './podcast-totals.selectors';
+import { MetricsType, METRICSTYPE_DOWNLOADS } from '../models';
+
+export const selectDownloadsSummaryTotal = createSelector(
+  selectMetricsTypeRoute,
+  selectPodcastMetricsFilteredTotal,
+  selectRoutedPodcastRanksTotalDownloads,
+  (metricsType: MetricsType,
+   downloadsTotal: number,
+   ranksTotal: number) => {
+    if (metricsType === METRICSTYPE_DOWNLOADS) {
+      return downloadsTotal;
+    } else {
+      return ranksTotal;
+    }
+  });
