+++ src/app/ngrx/reducers/selectors/index.ts
@@ -9,6 +9,7 @@ export * from './episode-metrics.selectors';
 export * from './episode-performance-metrics.selectors';
 export * from './downloads-chart.selectors';
 export * from './downloads-table.selectors';
+export * from './downloads-summary.selectors';
 export * from './loading.selectors';
 export * from './loaded.selectors';
 export * from './error.selectors';
